+++ src/editor/core/draw/particle/previewer/Previewer.ts
@@ -1,4 +1,5 @@
 import { EDITOR_PREFIX } from '../../../../dataset/constant/Editor'
+import { EditorMode } from '../../../../dataset/enum/Editor'
 import { IEditorOption } from '../../../../interface/Editor'
 import { IElement, IElementPosition } from '../../../../interface/Element'
 import { EventBusMap } from '../../../../interface/EventBus'
@@ -71,6 +72,7 @@ export class Previewer {
     element: IElement,
     position: IElementPosition | null = null
   ): { x: number; y: number } {
+    const { scale } = this.options
     let x = 0
     let y = 0
     const height = this.draw.getHeight()
@@ -79,8 +81,8 @@ export class Previewer {
     const preY = pageNo * (height + pageGap)
     // 优先使用浮动位置
     if (element.imgFloatPosition) {
-      x = element.imgFloatPosition.x!
-      y = element.imgFloatPosition.y + preY
+      x = element.imgFloatPosition.x! * scale
+      y = element.imgFloatPosition.y * scale + preY
     } else if (position) {
       const {
         coordinate: {
@@ -430,9 +432,14 @@ export class Previewer {
 
   public render() {
     // 图片工具配置禁用又非设计模式时不渲染
+    const mode = this.draw.getMode()
     if (
       !this.curElement ||
-      (this.curElement.imgToolDisabled && !this.draw.isDesignMode())
+      (this.curElement.imgToolDisabled && !this.draw.isDesignMode()) ||
+      (mode === EditorMode.PRINT &&
+        this.options.modeRule[EditorMode.PRINT]?.imagePreviewerDisabled) ||
+      (mode === EditorMode.READONLY &&
+        this.options.modeRule[EditorMode.READONLY]?.imagePreviewerDisabled)
     ) {
       return
     }
@@ -446,7 +453,16 @@ export class Previewer {
     options: IPreviewerDrawOption = {}
   ) {
     // 图片工具配置禁用又非设计模式时不渲染
-    if (element.imgToolDisabled && !this.draw.isDesignMode()) return
+    const mode = this.draw.getMode()
+    if (
+      (element.imgToolDisabled && !this.draw.isDesignMode()) ||
+      (mode === EditorMode.PRINT &&
+        this.options.modeRule[EditorMode.PRINT]?.imagePreviewerDisabled) ||
+      (mode === EditorMode.READONLY &&
+        this.options.modeRule[EditorMode.READONLY]?.imagePreviewerDisabled)
+    ) {
+      return
+    }
     // 缓存配置
     this.previewerDrawOption = options
     this.curElementSrc = element[options.srcKey || 'value'] || ''