+++ src/editor/core/draw/particle/table/TableOperate.ts
@@ -356,8 +356,8 @@ export class TableOperate {
     const trList = element.trList!
     const curTr = trList[trIndex!]
     const curTdRowIndex = curTr.tdList[tdIndex!].rowIndex!
-    // 如果是最后一行，直接删除整个表格
-    if (trList.length <= 1) {
+    // 如果是最后一行，直接删除整个表格（如果是拆分表格按照正常逻辑走）
+    if (trList.length <= 1 && element.pagingIndex === 0) {
       this.deleteTable()
       return
     }
@@ -457,8 +457,25 @@ export class TableOperate {
     const positionContext = this.position.getPositionContext()
     if (!positionContext.isTable) return
     const originalElementList = this.draw.getOriginalElementList()
-    originalElementList.splice(positionContext.index!, 1)
-    const curIndex = positionContext.index! - 1
+    const tableElement = originalElementList[positionContext.index!]
+    // 需要删除的表格数量（拆分表格）及位置
+    let deleteCount = 1
+    let deleteStartIndex = positionContext.index!
+    if (tableElement.pagingId) {
+      // 开始删除的下标位置
+      deleteStartIndex = positionContext.index! - tableElement.pagingIndex!
+      // 计算删除的表格数量
+      for (let i = deleteStartIndex + 1; i < originalElementList.length; i++) {
+        if (originalElementList[i].pagingId === tableElement.pagingId) {
+          deleteCount++
+        } else {
+          break
+        }
+      }
+    }
+    // 删除
+    originalElementList.splice(deleteStartIndex, deleteCount)
+    const curIndex = deleteStartIndex - 1
     this.position.setPositionContext({
       isTable: false,
       index: curIndex