+++ src/editor/interface/Editor.ts
@@ -75,6 +75,7 @@ export interface IEditorOption {
   letterClass?: string[]
   contextMenuDisableKeys?: string[]
   scrollContainerSelector?: string
+  pageOuterSelectionDisable?: boolean
   wordBreak?: WordBreak
   table?: ITableOption
   header?: IHeader
@@ -96,6 +97,7 @@ export interface IEditorOption {
   lineNumber?: ILineNumberOption
   pageBorder?: IPageBorderOption
   badge?: IBadgeOption
+  modeRule?: IModeRule
 }
 
 export interface IEditorResult {
@@ -131,4 +133,24 @@ export interface ISetValueOption {
 
 export interface IFocusOption {
   position?: LocationPosition
+  rowNo?: number
+  isMoveCursorToVisible?: boolean
+}
+
+export interface IPrintModeRule {
+  imagePreviewerDisabled?: boolean
+}
+
+export interface IReadonlyModeRule {
+  imagePreviewerDisabled?: boolean
+}
+
+export interface IFormModeRule {
+  controlDeletableDisabled?: boolean
+}
+
+export interface IModeRule {
+  [EditorMode.PRINT]?: IPrintModeRule
+  [EditorMode.READONLY]?: IReadonlyModeRule
+  [EditorMode.FORM]?: IFormModeRule
 }