+++ src/editor/utils/element.ts
@@ -1775,12 +1775,14 @@ export function getNonHideElementIndex(
   index: number,
   position: LocationPosition = LocationPosition.BEFORE
 ) {
-  if (!elementList[index]?.control?.hide) return index
+  if (!elementList[index]?.control?.hide && !elementList[index]?.area?.hide) {
+    return index
+  }
   let i = index
   if (position === LocationPosition.BEFORE) {
     i = index - 1
     while (i > 0) {
-      if (!elementList[i]?.control?.hide) {
+      if (!elementList[i]?.control?.hide && !elementList[i]?.area?.hide) {
         return i
       }
       i--
@@ -1788,7 +1790,7 @@ export function getNonHideElementIndex(
   } else {
     i = index + 1
     while (i < elementList.length) {
-      if (!elementList[i]?.control?.hide) {
+      if (!elementList[i]?.control?.hide && !elementList[i]?.area?.hide) {
         return i
       }
       i++